Updated README.md
Just in case, in Weechat, plugins and scripts are different in definitions. The current project seems to be a script for plugin Python for Weechat.

In addition, since release of Weechat v3.2, XDG directories are prioritized and hence the location mentioned in the Readme may not present the current state.